post

+++ app/Http/Controllers/TaskController.php
@@ -17,11 +17,21 @@ class TaskController extends Controller
      */
     public function index()
     {
-           $tasks = Task::all();
+              $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            $data += $this->counts($user);
+            return view('tasks.index', $data);
+        }else {
+            return view('welcome');
+        }
+
         //
     }
 
@@ -49,12 +59,14 @@ class TaskController extends Controller
     {
         $this->validate($request,[
             'status' =>'required|max:10',
+            'content' => 'required|max:191',
             ]);
-            
-        $task = new Task;
-        $task->status = $request ->status;
-        $task->content = $request->content;
-        $task->save();
+        
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->content
+        ]);
+        
 
         return redirect('/');
         //
@@ -121,10 +133,14 @@ class TaskController extends Controller
      */
     public function destroy($id)
     {
-         $task = Task::find($id);
-        $task->delete();
+        
+        $task = \App\Task::find($id);
 
-        return redirect('/');
-        //
+        if (\Auth::user()->id === $task->user_id) {
+            $task->delete();
+        }
+
+        return redirect()->back();
+
     }
 }
